@@ -17,14 +17,14 @@
         }
     ],
     "require": {
-        "php": "^8.0",
+        "php": "^8.1",
         "ext-mbstring": "*",
         "cocur/slugify": "^4.0",
         "mrclay/minify": "^3.0",
         "nesbot/carbon": "^2.41",
         "psr/container": "^2",
         "psr/log": "^3",
-        "qubus/exception": "^1.2",
+        "qubus/exception": "^2.0",
         "symfony/var-dumper": "^5.3"
     },
     "require-dev": {
